CentralNotice Banner Bucketing [FR2012 626]
Campaigns now have the concept of banner buckets. This will
allow for A/B testing of banners.

Was going to remove Special:BannerListLoader but apparently
it's faster than an API call so it's back! -- and I left
the API call because why not.

Patchset 8: Making sure BannerListLoader is localized
Patchset 9: Adding back in bucketing after PS6/7? removed it

Patchset 11: Simplify interface, all filtering is done in the API

Patchset 15: rebase onto BannerChooser
* explode Special:BannerAllocations by target (anonymous/bucket).
TODO - needs to be usable and sane

Change-Id: I95cc31a089e8ccd4fa12ff928ee8f50b3e38ea45