CentralNotice Banner Bucketing [FR2012 626]
Campaigns now have the concept of banner buckets. This will
allow for A/B testing of banners.

Was going to remove Special:BannerListLoader but apparently
it's faster than an API call so it's back! -- and I left
the API call because why not.

Patchset 8: Making sure BannerListLoader is localized
Patchset 9: Adding back in bucketing after PS6/7? removed it

Patchset 11: Simplify interface, all filtering is done in the API

Patchset 15: rebase onto BannerChooser
* explode Special:BannerAllocations by target (anonymous/bucket).
TODO - needs to be usable and sane

==== adamw broke patch.  continuing...

Patchset 1: === Patchset 16 above

Change-Id: I36be97e850ce79aadab0cda7bb39d493078217dc