Fixing Multiselector Suckage
The JS that makes the fancy multiselector was running before the CSS
that gave it a good width. So now we wait for 250ms. Obviously this
is a terrible hack -- but it'll do until I rewrite the thing.

Change-Id: Ic8333f7125c6420a89c8a84618be871d2cdbb134
(cherry picked from commit 05dcf72431808be7931aa6d76d71b24318d78e13)