(Bug 51465) No Longer Filtering Mobile Devices
MobileFrontend leaves a cookie behind when switching back and
forth from mobile view on a desktop browser. This is impeding
people viewing banners relating to the VE rollout.

Removing the filtering comes at the expense of potentially
exposing users to banners that don't work particularly well on
small screens. This is judged as a reasonable risk since those
users are making a concious effort to choose to view a potentially
limited site and banners typically do not affect functionality.

Change-Id: I3c2421aee4a1f3696dcbbcefc95f4f0e6657a7a4
Bug: 51465