Correct Banner Over Allocation
When we distribute the remaining slots, because we do not check if
we have over allocated, we could end up with a situation where a
banner that did not need any extra slots got one.

E.g. 3 banners A, B, and C with weight 5, 100, and 100 would allocate
A: 6.66%
B: 46.66%
C: 46.66%

Now they will allocate
A: 3.33%
B: 50.00%
C: 46.47%

Change-Id: I70c6258b52bd1656d0967e45fc4dd74175ee6a57