(FR #1048) Campaign throttle knob
This patch contains the support and UI for the feature, but does not affect
allocations.

Campaign throttle allows CentralNotice admins to limit the traffic going to
their banners to a percentage of the total potential maximum allocation.
Leftover allocations are rolled over to campaigns at the next lower priority
levels.

Change-Id: I04d81df646c62580b5337b90f32db04c2ca37f59