Make the updater work
No idea what patch actually makes this work in master; but this
will make it work here. (We moved the hook definition around a
lot so tracing history is hard. However; compare to what is currently
in master:
https://git.wikimedia.org/blob/mediawiki%2Fextensions%2FCentralNotice/1a25675c5bd2daef9ddbe1c1fcac1127713a90b9/CentralNotice.hooks.php#L31

The reason the unit tests are failing because before 1.22 we didn't have a working sqlite schema -- and I didn't backport that.

Change-Id: I8835636ce7724e1802f78eca361821f80ca76503