Try geoiplookup.wikimedia.org only if client used IPv6
Currently, CentralNotice will fire a request to geoiplookup.wikimedia.org if
bits.wikimedia.org/geoiplookup failed to geolocate the client IP.
geoiplookup.wikimedia.org is IPv4-only, so the thought is that dual-stack users
would be forced to connect via IPv4, allowing them to be geolocated. But IPv6
clients are just one reason lookups can fail. The response from bits may be
blank if the client connected via IPv4 using an IP address that was not
familiar to the MaxMind database. In such cases, there is no point in making a
secondary request to geoiplookup.wikimedia.org, since it will only produce the
same result. This patch ensures that the secondary request to
geoiplookup.wikimedia.org is only made when it could plausibly be useful --
i.e., when the address family of the client IP for the initial lookup attempt
was IPv6.

Change-Id: I18e7dcfc3ae7bbeb70118f63e1f7cb86bcdcd535