Harmonize format of GeoIP cookie
When synthesizing the GeoIP cookie, it got prefixed by “GeoIP=”, if
window.Geo.country was not set. And it did not get this prefix, if
window.Geo.country was set.

We now drop the unneeded “GeoIP=” prefix altogether to make sure the
format matches in both cases. Thereby, we match the regular expression
used for parsing of window.Geo in both cases.

Change-Id: Iff5c2aa50c7fad0cee44290ead015b8cf34c0645