Add dates to banner-hiding cookies (client side)
This change allows the bannerController to understand both
the old 'hide' value and a new JSON format:
{
	"v" : 1, (cookie version)
	"created" : 12345678,
	"reason" : "close"
}
where "created" is in seconds since the epoch, and "reason" indicates
why the banner should be hidden (i.e., the user has donated or has
clicked the close button on a banner.)

An array of cookie durations is provided in mw.config corresponding to
the possible hide reasons.

This should be thoroughly deployed before updating Special:HideBanners
to write cookies in the new format.

Change-Id: I02988b260d50db2bc61fe32f95490f0339a52e38