Add dates to banner-hiding cookies (server side)
This should not be deployed until the JS portion is everywhere.

Update Special:HideBanners to change the value of hiding cookies
from 'hide' to a new JSON format:
{
	"v" : 1, (cookie version)
	"created" : 12345678,
	"reason" : "close"
}
where "created" is in seconds since the epoch, and "reason" indicates
why the banner should be hidden (i.e., the user has donated or has
clicked the close button on a banner.)

Not including special cases for requests from the old JS, as the old
JS is getting a zero value for $wgNoticeCookiesShortExpiry and thus
sending 0 for duration.  Also, no more old JS requests have been caught
in the log filter in the past week.

Change-Id: I62834899c2a861099340003c5528440595e67e6c