Pager for campaigns on campaigns and banner pages
Refactor as a TablePager existing code for displaying a list of
campaings, then use it on the banner edit page for showing
campaigns associated with a banner.

Note: this change removes editing via the list of campaings, if there
are more than 5000 campaigns. This could be fixed by refactoring
part of CentralNotice (the PHP class).

Change-Id: I48e651690980d4b1d6a2ec37f584f4cbe8ad5c68