Preload JS comes with no baggage
It's just a snippet that runs before the banner is inserted.  There is no
mandatory interface such as implementing the alterImpressionData hook or
returning a boolean.

A new variable is introduced, impressionData.result, which is checked by
the banner controller.  If preload js has set this variable to 'hide',
no HTML is inserted into the DOM.

Bug: T86100
Change-Id: I8ffcbb44ac15cbb91dad6786cc2055f589174d1d