Check choiceData for staleness
There may be caching issues that we can't control that cause
some users to get stale choiceData. Here we test for by checking
if we got any campaings that should have ended, and turn off
campaigns completely if we find it. A 15-minute leeway is also
built in to allow for cache refresh time.

Bug: T88276
Change-Id: Ib0f2ede85083d9568190f0211bd80f68e745f171