Large banner limit mixin
In campaigns with this mixin enabled, a reader will see a large
banner only once, then will only see small banners for a configurable
period.

Requires a 4-bucket campaign with large banners in A & B and small
banners in C & D.

Transitions readers from bucket A to C and B to D, or optionally to
either of C and D randomly.

Cookie name is configurable, but defaults to same as on-wiki scripts
for compatibility.

TODO: after on-wiki scripts are no longer in use, move indicator to
localStorage when available.

Bug: T90919
Change-Id: Idac967fc2c4edb1d35350a2c9611c53c1d5a8d36