impressionDiet: handle deployed problem for campaign categories using legacy
An error was deployed to WMF production that caused fundraising campaigns to move
to using the KV store when they weren't supposed to (because they used the category
'fundraising' rather than 'Fundraising').

To remedy this, we fix the default value of $wgCentralNoticeCategoriesUsingLegacy
and keep clients that already have a KV store item for a given identifier
on the KV store system. This should keep impression counts consistent. For those
cases, a reverse migration back to cookies is not deemed necessary.

Change-Id: I75017856505d289941a690309c7cdeae8c2909ba