impressionDiet: don't count impressions if banner hidden by close cookie
TODO: This is a temporary fix. It addresses an issue currently affecting fundraising
campaigns, but does not help the bigger problem of order of execution of mixin hooks.
Also, there are probably similar unresolved issues with the interactions of the impression
diet mixin and the large banner limit mixin (currently not enabled in FR campaigns).

Bug: T120698
Change-Id: I877f011790a8e366521e15548728a24f77ed2468