Improve impression diet state machine
This fixes two bugs which were causing extra waitdate hides.
* Even when readers have a waitUntil set, increasing the campaign maximum
impressions seen per cycle should result in the additional banners.
* Got rid of an unnecessary error condition.  It's clear that every code
path in the main switch will assign a value to the hide variable.

Also gets rid of the waitnorestart state.  This was equivalent to waitdate,
just check the single-shot campaign configuration to tell if waitdate is
going to restart or not.

Bug: T121178
Change-Id: I7d1c2b6672d5ef12405d1f43a36ea3f5e01da21c