Account for DB lag when refreshing cached ChoiceData
This fixes an apparent race condition in which ObjectCache-cached ChoiceData
is invalidated following a change in campaign or banner settings, then a
client request has a cache miss and re-sets the cache using data form a laggy
DB slave, causing stale ChoiceData to be cached and served. For details about
the solution, see the inline documentation for
WANObjectCache::getWithSetCallback(), in MediaWiki core.

Bug: T144393
Change-Id: Id4e46b653554d39ac5721c4327774ccfce63b734