CNChoiceDataModule: Avoid deprecated getModifiedHash()
ResourceLoaderModule::getModifiedHash was deprecated in 1.26
in favour of appending information to the array from
getDefinitionSummary (aded in MediaWiki 1.23).

The main benefit is so that getVersionHash() does not trigger
multiple phases of serialize, json_encode, md5 and/or sha1.
Instead, it only happens once and outside the concern of module
sub classes.

This is the last use of getModifiedHash() in Wikimedia Git,
Core and other extensions have also adopted getDefinitionSummary
(See Wikibase, Cite, ULS, EventLogging, etc.).

Change-Id: I0f76af33dc6f309d8c1969b078706e0283608a29