Emit CSP headers on banner previews
FIXME: as soon as I80f6f469ba4c0b60 is available in core, get rid of
$wgCentralNoticeContentSecurityPolicy and use their stuff.

To test, add this to your localsettings:
$wgCentralNoticeContentSecurityPolicy =
  "default-src data: blob: 'unsafe-inline' 'unsafe-eval' 'self';";

Then create a banner that tries to load an external resouce, and
preview it with the ?banner=<banner_name> parameter.

Note: works with any new browser except Firefox 58 and below.
In FF 59+ you need to set security.csp.enable_violation_events to
true in about:config

https://developer.mozilla.org/en-US/docs/Web/API/SecurityPolicyViolationEvent#compatNote_1

Bug: T190100
Change-Id: Ie34dc8014e306883a3add07c8fe9e39c8486c603