Banner live preview
This is yet a very first approach on adding a live-preview feature to
banner editing screen.

At the moment it automatically displays a banner preview in a separate
fieldset above the content editing area and updates it when content is
updated or related messages texts (if any) are updated in the form.

Right now it's doing a very simple job on replacing embed messages,
that part probably needs to be improved.

Bug: T208125
Change-Id: Ib8cbc2d7bee2978819f0d7fc2a01381d142bcd02