ext.centralNotice.geoIP: Remove redundant variable, and use then()
Use then() instead of done() so that the Cookie IO is scheduled
for an async tick, instead of synchronously in the same tick
that fulfilled the geoPromise.

This is better for performance (less long back-to-back chunks of
execution blocking the browser thread), and makes the code
compatible with standard JS Promise (which doesn't have a
synchronous "done" method, probably a good thing).

Change-Id: Ic583fa5ade6307503a4ff596840eccb87426beae