Banner: Use self to reference class name in PHP and PHPDoc
Use self() instead of Banner() [class name] for safety. Take
for example, the name of this class changing in the future,
which could be very unlikely but if it does, having self() in
the code base will safe us the time and risk to go changing
all references to this class in the codebase and if one forgets
to do this, it can break a lot of things. So, to avoid this risk,
having self() instead of Banner() is better.

Change-Id: I7492fa31cda53f2e8e6f7394574d4c36a913b8b6