ext.centralNotice.startUp: Convert to packageFiles module
Give this module its own directory and register it as a ResourceLoaderFileModule
through the 'packageFiles' option instead of the 'scripts' option.

On its own this changes nothing, but for future changes it two things.

1) If the file were to contain multiple files in the future then, unlike
old-style modules, these files will not be concatenated together and executed
as one. Instead, with 'packageFiles' the additional files will be exported
by name, and then the exports are concatenated and executed on the client
as one.

The main difference being that it enables use of require() for relative
imports. For example, 'Foo.js' can export a class Foo, and this can be
imported as `var Foo = require('./Foo.js')` without requiring some global
variable to stash it under.

2) It allows MediaWiki to easily augment the module with extra data in
the form of a virtual file. For example './data.json'. This is easier
to manage than prepending generated JS statements in PHP that e.g.
stash stuff under mw.config among other ones as globals that could
clash and are hard to track in terms of dependency tree.

This is prep for solving T219342#5068410. Although it could be
worked around if needed.

Bug: T193826
Change-Id: I6bb168c26fab717772567045d46a4fd448231f84