ext.centralNotice.display: Convert to using packageFiles
Also give the module bundle its own directory while at it (per T193826).

This is in preparation for reducing the CN overhead on the RL startup
module. Specifically, to no longer export the CN-related config data
through 'mw.config' – This blocks rendering of pages, and delays
the point where we can start fetching modules (including CN banners).

For T219342, these exports will be moved to instead be part of the
module that needs it which means 1) we only load it when needed, and
2) it will load in parallel with other modules, instead of blocking
the point where we can start loading modules.

Change-Id: I606d227e6b34c509ddeadbce0e4f49821633a807