Remove 'wgCentralNoticeCategoriesUsingLegacy' from page view HTML
Follows-up b3ff2cf5ba, which did this for wgCentralNoticeCookiesToDelete

This is needed for ext.centralNotice.display. Bundle it with
that module instead. This means it will only load when needed,
and will load in parallel with other modules, instead of where it
was before, where it is a blocking part of the HTML download
for rendering articles, and also delaying when we can start
downloading modules (including CN banners). Win-win :)

Bug: T219342
Change-Id: Ie56c67fed25dd445291053c8ce14c2ab5b3d0dd5