Campaign fallback
The change allows for campaigns to fallback. Now it will not stop after the first campaign
fails to display a banner and instead will iterate through the rest of
matching campaigns trying to find another candidate for display. The fallback will not
occur if a banner was not shown due to a campaign configuration issue.

- The change does not contain any additions to banner allocation preview
- The `campaignStatuses` property was added into the list of data being submitted during
the impression recording. The property contains all the campaigns were available for choosing
and their statuses at the end of the choosing process. This makes it possible to have better
insights on failed campaigns.
- `wgCentralNoticeMaxIterations` configuration variable is added to set the limit for number
of iterations could be performed during campaign fallback/choosing procedure.

Bug: T124969
Change-Id: I2584419c65a169e5a23f3ff84eae89547cfeac4a