Only register adminUI modules on the CN infra wiki (e.g. Meta-Wiki)
These modules are only loaded from the various CentralNotice-related
special pages, which in turn are only registered by
CentralNoticeHooks::onRegistration on wikis where
wgNoticeInfrastructure is `true` (such as on Meta-Wiki in WMF prod).

Move the registration of these RL modules to the same place where
these special pages are registered.

The motivation for this change is that CentralNotice currently
registers 1.2 KB of module metadata on all Wikipedia page views,
of which most is for adminUI modules that are never loaded there.

Bug: T221805
Change-Id: I8db27564a977491c99a8fcecdd90512c1abc59c0