tests: Don't write to the actual database
setUpBeforeClass is called *before* the mock db is setup, hence all writes
will happen on the actual database. This must be avoided at any cost.

Since all writes are now happening on the fake DB, it's also guaranteed
that the banner won't exist before tests are executed, hence it's enough
to delete it in tearDown.

Change-Id: Ie30d84ff5bdcaf64559f16b3452e71e7d1f6429c