Only add tabs to special pages
Following Ia867166221, the hook to add tabs for the CN admin
interface runs on all pages, not just special ones, causing the
tabs to be added on a page called "CentralNotice" and subpages
thereof. This patch adds a check to prevent that.

Bug: T311944
(cherry-picked from a2151186441b10820e2e068499686dc70f25f005)

Change-Id: I370c3307e6a3e27f2496e0e15da5dcbca68e9f58