styles: Remove unnecessary 'mediawiki.ui/variables' import
Removing 'mediawiki.ui/variables.less' `@import` as the vars are
not used at all. We could also update to new standard
'mediawiki.skin.variables.less' but given that it's just this single
file and we'd have to bump the extension dependency to v1.41.0, it
seems more sensible to just remove the import.

Bug: T332541
Change-Id: I06c9e4c24883d5a570b03ca7287f3f3f0e6c0c27