Move archived campaigns to a new subaction of Special:CentralNotice
The current behaviour of the Special:CentralNotice admin interface is to always
load all campaigns, and hide archived ones with CSS. At Wikimedia this results
in loading HTML for literally thousands of campaigns, which in almost all cases
are not needed. This has produced a noticeable slowdown in the admin interface,
and made it difficult or impossible to use on mobile.

This patch makes Special:CentralNotice only load unarchived campaigns, and adds
a new subaction=showArchived to show archived ones when needed. It removes the
now unnecessary CSS and JS checkbox filter.

The lists on banner detail pages of "Campaigns using this banner" will now show
both archived and unarchived campaigns.

Bug: T328424
Change-Id: I9da7fd8fdd6ecd7f7ad901d8b30c34ca3b58eeb9