Stop binding against core's ILocalizedException interface
This code was added via Ifbdad6c a few years ago.

As far as I can tell there was never really a point in using the core
interface. Maybe convenience? Or just having a pattern to follow? As
it turned out that pattern was not so great, see T287405.

There is apparently only a single exception that works like this.
Core does have a few more that implement the ILocalizedException
interface, but these are not relevant here.

Bug: T287405
Change-Id: Ice8cd270fc4be2f257955dc75af1a10ffe77564c