Drop misleading boolean comparisons with "1"
While these are not wrong I find them confusing. The database fields
are TINYINT(1), not strings. They are meant to store booleans,
i.e. only 0 and 1 can be stored. Comparisons with PHP's relaxed ==
operator can behave weird for historical reasons. The actual
numerical value doesn't matter, only if it's something that resembles
true or false. I find it more straightforward to just cast the numbers
0 and 1 to bool in such a situation. Most code that works with these
fields already does the same.

Note that the confusing abbreviation "not" here is not the word "not"
but means "notice".

Change-Id: I027a35457da2316a438e7fc717180248f053b882