Remove misleading "protected" visibility of private code
"Protected" means "public to subclasses". But most of this code
doesn't even have subclasses, and probably never should have one.
There should be no reason to make such implementation details
public. If there really is a reason to do this it's very easy to
set the visibility of the method in question back to "public" or
"protected", if needed.

Change-Id: Ie65732bf8590ba14b8b6fe429c02a72474df9b26