Replace more Xml::… methods with Html::… equivalents
This is the third step after I3617627 and If1f42c4. The replacements
are all very trivial because the method signatures are almost
identical. Ony difference is that the Html class doesn't like null
for the attributes but expects an empty array.

Only form-related stuff is left after this patch.

Bug: T341775
Change-Id: I86fb0639ac90b37fb5959b88d6066075e75b78f9