Use mw.util.messageBox
Instead of building the HTML for the message box and hardcoding the
styles, use the core function. Also keep track of the error box
elements in the error state tracker for removal when the error is
cleared.

Testing notes: Create a new campaign at Special:CentralNotice and
enable the 'Impression diet' mixin. Then try various values in one
of the integer fields like 'Maximum impressions'. The error alert
should show up while the field is invalid (and the submit button
should be disabled), and the error alert should go away when the
field is value (and the submit button should be re-enabled).

Bug: T375902
Change-Id: I3e7554abd88c77017381c9949732173ca7ac6e1f