Use Title::makeTitle in Banner::getTitle
In this class the dbKey never contains a namespace, so it is not needed
to use Title::newFromText. The banner name is validated via
Banner::isValidBannerName to use ascii letters and so it is not
necessary to use Title::makeTitleSafe, as most callers of getTitle()
does not expects to get a null. So use Title::makeTitle instead.

For extra safety limit the length of the banner name to 230 bytes
(database allows 255, but there is prefix added for the page in
mediawiki namespace)

Change-Id: Iaa729c224492a36bce91d71a65c26e3ed118d2f2