Do not show banners when custom site scripts are disallowed
Since banners can contain Javascript, do not display them when
user-controlled scripts are disallowed (e.g.
OutputPage::disallowUserJs() has been used).

In practice this is mostly a no-op since disallowUserJs is only
used on special pages and banners are already not shown there,
but better to be consistent.
Also, disabling for special pages is explained as a UX fix, and
it's good to have the security rationale explicitly mentioned
in the code.

Code copied from Hooks::addTour() in GuidedTour.

Bug: T373737
Bug: T373738
Change-Id: If9be080389d2ec64e1c49981c11241f196a9795e