Remove unused wgCentralSelectedMobileBannerDispatcher
== Background ==

This was added in 7d1c58482 (I4abc03e3fdb1), but never declared
in extension.json, and never set in production in wmf-config. It
seems the discussion at T158030 settled on not needing it since
the banner content indeed doesn't vary between desktop and mobile.

Note that impression metrics are sent to /beacon/impression,
which is configured separately (wgCentralBannerRecorder).

While at it, also remove an attempt at documenting the config var
from the README ("wgCentralMobileSelectedBannerDispatcher"), which
uses a different word order and doesn't exist, but was presumably
intended to describe wgCentralSelectedMobileBannerDispatcher.

== Why now ==

As part of T390929, I'm auditing code that makes calls to
MobileContext::shouldDisplayMobileView outside index.php or api.php.

This hook is called while building the mw.config dataset in
load.php?modules=mediawiki.base. This appears to be one of very few
places where that variation exists, and so as part of m-dot sunsetting,
I'd like to explore not having to introduce Vary:X-Subdomain on
load.php responses.

Bug: T154954
Bug: T158030
Bug: T400586
Change-Id: Ieb6e2523a752b8f8aedcbb793c4a84ad00565445