Remove shouldDisplayMobileView/getMobileUrl transform of wgCentralBannerRecorder
After I8eeedb098c79 in wmf-config, this is "/beacon/impression" instead of the
almost-full-URL-but-protocol-relative "//en.wikipedia.org/beacon/impression".

The JavaScript code at resources/ext.centralNotice.display/index.js
parse and this with `new URL` and passes it to sendBeacon or <img src>,
all of which support relative URLs, that thus naturally connect to
the current domain.

Bug: T400586
Change-Id: I657e34c6cbfc16f3f464aac852b5845195789899